Use virtualedit=onemore
